remove some deprecated patterns

+++ hardware/chisel/src/test/scala/unittest/AluTest.scala
@@ -53,7 +53,7 @@ class TestAluVector(c: AluVector) extends PeekPokeTester(c) {
       // HLS shift left by >> negative number
       // b always < 0 when opcode == 4
       for (i <- 0 until size) {
-        res(i) = a(i) << ((-1*b(i)) & mask)
+        res(i) = a(i) << ((-1*b(i)) & mask).toInt
       }
     } else {
       // default
